Backend/requirements.txt
#Lists Python libraries you must install.
fastapi
uvicorn[standard]
beanie
motor
pydantic
python-dotenv
